updated code with latest dev changes and renamed to blackduck_component_risk
