+++ dojo/fixtures/test_type.json
@@ -538,7 +538,7 @@
 	},
 	{
 		"fields": {
-			"name": "Blackduck Hub Scan V2"
+			"name": "Blackduck Component Risk"
 		},
 		"model": "dojo.test_type",
 		"pk": 169
