+++ dojo/forms.py
@@ -343,7 +343,7 @@ class ImportScanForm(forms.Form):
                          ("Twistlock Image Scan", "Twistlock Image Scan"),
                          ("Kiuwan Scan", "Kiuwan Scan"),
                          ("Blackduck Hub Scan", "Blackduck Hub Scan"),
-                         ("Blackduck Hub Scan V2", "Blackduck Hub Scan V2"),
+                         ("Blackduck Component Risk", "Blackduck Component Risk"),
                          ("Openscap Vulnerability Scan", "Openscap Vulnerability Scan"),
                          ("Wapiti Scan", "Wapiti Scan"),
                          ("Immuniweb Scan", "Immuniweb Scan"),
