+++ dojo/tools/factory.py
@@ -65,7 +65,7 @@ from dojo.tools.testssl.parser import TestsslCSVParser
 from dojo.tools.hadolint.parser import HadolintParser
 from dojo.tools import SCAN_SONARQUBE_API
 from dojo.tools.aqua.parser import AquaJSONParser
-from dojo.tools.blackduck_v2.parser import BlackduckHubParser
+from dojo.tools.blackduck_component_risk.parser import BlackduckHubParser
 from dojo.tools.h1.parser import HackerOneJSONParser
 from dojo.tools.xanitizer.parser import XanitizerXMLParser
 from dojo.tools.trivy.parser import TrivyParser
@@ -190,7 +190,7 @@ def import_parser_factory(file, test, active, verified, scan_type=None):
         parser = KiuwanCSVParser(file, test)
     elif scan_type == 'Blackduck Hub Scan':
         parser = BlackduckHubCSVParser(file, test)
-    elif scan_type == 'Blackduck Hub Scan V2':
+    elif scan_type == 'Blackduck Component Risk':
         parser = BlackduckHubParser(file, test)
     elif scan_type == 'Sonatype Application Scan':
         parser = SonatypeJSONParser(file, test)
